backend/requirements.txt
asgiref==3.8.1
asttokens==2.4.1
bcrypt==4.1.3
build==1.2.1
cffi==1.16.0
click==8.1.7
colorama==0.4.6
comm==0.2.2
cramjam==2.8.3
cryptography==42.0.8
debugpy==1.8.1
decorator==5.1.1
Django==5.0.6
django-cors-headers==4.4.0
django-filter==24.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dnspython==2.6.1
executing==2.0.1
ipykernel==6.29.4
ipython==8.25.0
jedi==0.19.1
joblib==1.4.2
jupyter_client==8.6.2
jupyter_core==5.7.2
Markdown==3.6
matplotlib-inline==0.1.7
mysqlclient==2.2.4
nest-asyncio==1.6.0
numpy==2.0.0
packaging==24.1
pandas==2.2.2
parso==0.8.4
pip-tools==7.4.1
platformdirs==4.2.2
prompt_toolkit==3.0.47
psutil==6.0.0
pure-eval==0.2.2
pycparser==2.22
Pygments==2.18.0
PyJWT==2.8.0
pymongo==4.7.3
pyproject_hooks==1.1.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-snappy==0.7.2
pytz==2024.1
pywin32==306
pyzmq==26.0.3
scikit-learn==1.5.0
scipy==1.13.1
setuptools==70.1.0
six==1.16.0
sqlparse==0.5.0
stack-data==0.6.3
threadpoolctl==3.5.0
tornado==6.4.1
traitlets==5.14.3
tzdata==2024.1
wcwidth==0.2.13
wheel==0.43.0
winkerberos==0.10.0
